+++ apps/docs/.dumirc.ts
@@ -9,6 +9,11 @@ export default defineConfig({
             target: "http://localhost:8888/",
             changeOrigin: true,
             pathRewrite: { "^/api": "" }
+        },
+        "/dcau-api": {
+            target: "http://172.22.50.33:20085",
+            changeOrigin: true,
+            pathRewrite: { "^/dcau-api": "/dcau-api" }
         }
     }
 })