+++ packages/upload-sdk/src/index.ts
@@ -1,4 +1,5 @@
 export { UploadManager } from "./core"
 export { useUpload } from "./hooks/useUpload"
+export { UploadBatchError } from "./utils"
 
 export * from "./types"