+++ packages/upload-sdk/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./formatError"
+export * from "./UploadBatchError"